Renomeando interfaces

+++ src/main/java/com/bank/balance/app/usecases/impl/SaveBalanceEntryHistory.java
@@ -1,17 +1,17 @@
 package com.bank.balance.app.usecases.impl;
 
 import com.bank.balance.app.repositories.IBalanceEntryDocumentRepository;
-import com.bank.balance.app.repositories.IBalanceEntryRepository;
+import com.bank.balance.app.repositories.IBalanceEntryEntityRepository;
 import com.bank.balance.app.usecases.ISaveBalanceEntryHistory;
 import org.springframework.stereotype.Service;
 
 @Service
 public class SaveBalanceEntryHistory implements ISaveBalanceEntryHistory {
 
-    private final IBalanceEntryRepository balanceEntryRepository;
+    private final IBalanceEntryEntityRepository balanceEntryRepository;
     private final IBalanceEntryDocumentRepository balanceEntryDocumentRepository;
 
-    public SaveBalanceEntryHistory(final IBalanceEntryRepository balanceEntryRepository, final IBalanceEntryDocumentRepository balanceEntryDocumentRepository) {
+    public SaveBalanceEntryHistory(final IBalanceEntryEntityRepository balanceEntryRepository, final IBalanceEntryDocumentRepository balanceEntryDocumentRepository) {
         this.balanceEntryRepository = balanceEntryRepository;
         this.balanceEntryDocumentRepository = balanceEntryDocumentRepository;
     }
@@ -19,7 +19,9 @@ public class SaveBalanceEntryHistory implements ISaveBalanceEntryHistory {
     @Override
     public void execute() {
         final var balanceEntries = balanceEntryRepository.findAllForArchives();
-        balanceEntryDocumentRepository.save(balanceEntries);
-        balanceEntryRepository.deleteAll(balanceEntries);
+        if (!balanceEntries.isEmpty()) {
+            balanceEntryDocumentRepository.save(balanceEntries);
+            balanceEntryRepository.deleteAll(balanceEntries);
+        }
     }
 }
